+++ src/api/swagger-types/SkillsController/getAllSkills.ts
