+++ src/api/swagger-types/UsersController/signup.ts
