+++ src/api/useSignupMutation.ts
@@ -1,11 +1,11 @@
+import { AxiosMutationOptions } from './generic/types/axios-mutation-options.type';
+import { useAxiosMutation } from './generic/useAxiosMutation';
 import {
   path,
   RequestBody,
   SignupError,
   SignupSuccess,
-} from './dynamic-types/UsersController/signup';
-import { AxiosMutationOptions } from './generic/types/axios-mutation-options.type';
-import { useAxiosMutation } from './generic/useAxiosMutation';
+} from './swagger-types/UsersController/signup';
 
 export const useSignupMutation = (
   options?: AxiosMutationOptions<SignupSuccess, SignupError, RequestBody>
