+++ src/api/useSkillsQuery.ts
@@ -1,9 +1,9 @@
-import { GetAllRolesError } from './dynamic-types/RolesController/getAllRoles';
+import { useAxiosQuery } from './generic/useAxiosQuery';
+import { GetAllRolesError } from './swagger-types/RolesController/getAllRoles';
 import {
   GetAllSkillsSuccess,
   path,
-} from './dynamic-types/SkillsController/getAllSkills';
-import { useAxiosQuery } from './generic/useAxiosQuery';
+} from './swagger-types/SkillsController/getAllSkills';
 
 export const useSkillsQuery = () =>
   useAxiosQuery<GetAllSkillsSuccess, GetAllRolesError>('skills', path, 'GET');
